+++ number-rogue/index.html
@@ -4,9 +4,16 @@
 		<meta charset="UTF-8" />
 		<link rel="icon" href="/favicon.png" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		<meta name="theme-color" content="#6272a4" />
-		<meta property="og:title" content="Number Rogue" />
-		<meta property="og:description" content="A calculating roguelike." />
+		<meta property="og:title" content="Calculate It" />
+		<meta property="og:description" content="A Calculate It clone." />
+		<link rel="preconnect" href="https://fonts.googleapis.com" />
+		<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+		<link rel="preconnect" href="https://fonts.googleapis.com" />
+		<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+		<link
+			href="https://fonts.googleapis.com/css2?family=JetBrains+Mono:ital,wght@0,100..800;1,100..800&family=Rubik:ital,wght@0,300..900;1,300..900&display=swap"
+			rel="stylesheet"
+		/>
 		<link rel="apple-touch-icon" href="/favicon.png" />
 		<!--
       manifest.json provides metadata used when your web app is installed on a
@@ -17,6 +24,6 @@
 	</head>
 	<body>
 		<div id="root"></div>
-		<script type="module" src="/src/number-rogue/src/main.tsx"></script>
+		<script type="module" src="/src/number-rogue/main.tsx"></script>
 	</body>
 </html>
