+++ package.json
@@ -11,6 +11,7 @@
     "sm-update": "git submodule update --init --recursive"
   },
   "dependencies": {
+    "canvas-confetti": "^1.9.3",
     "dexie": "^4.0.11",
     "dexie-react-hooks": "^1.1.7",
     "react": "^18.3.1",
@@ -22,6 +23,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.15.0",
+    "@types/canvas-confetti": "^1.9.0",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@types/seedrandom": "^3.0.8",
