+++ src/number-rogue/App.scss
@@ -1,0 +1,36 @@
+.container {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	max-width: 1000px;
+	width: 100%;
+	margin: 0 auto;
+	gap: 20px;
+	transition: width 0.3s ease, max-width 0.3s ease;
+
+	.fab {
+		position: fixed;
+		bottom: 20px;
+		right: 20px;
+		display: flex;
+		flex-direction: column-reverse;
+		gap: 10px;
+
+		button {
+			&:hover {
+				cursor: pointer;
+			}
+		}
+	}
+
+	@media (max-width: 600px) and (orientation: portrait) {
+		max-width: 100%;
+		width: 100%;
+
+		.fab {
+			bottom: 16px;
+			right: 16px;
+			gap: 8px;
+		}
+	}
+}
