+++ src/number-rogue/components/Calculator/Calculator.scss
@@ -5,6 +5,7 @@
 	width: fit-content;
 
 	.display {
+		font-family: 'Jetbrains Mono', monospace;
 		display: inline-block;
 		white-space: nowrap;
 		overflow-x: auto;
