+++ src/number-rogue/components/Calculator/calculator-config.ts
