+++ src/number-rogue/components/Scenes/GameScene.scss
@@ -1,0 +1,78 @@
+.game {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	padding: 60px 0;
+	gap: 20px;
+
+	.info {
+		display: flex;
+		flex-direction: column;
+		gap: 20px;
+
+		.seed {
+			display: flex;
+			align-items: center;
+			gap: 4px;
+
+			.seed-input {
+				background: #444;
+				color: #eee;
+				border: none;
+				font-size: 14pt;
+				border-radius: 4px;
+				padding: 4px 8px;
+				margin-bottom: 2px;
+
+				&:focus {
+					outline: none;
+				}
+			}
+
+			.reseed-button {
+				:hover {
+					cursor: pointer;
+				}
+			}
+		}
+
+		.spaced-out {
+			display: flex;
+			justify-content: space-between;
+			align-items: baseline;
+			width: 100%;
+		}
+
+		.money-text {
+			color: lime;
+			font-size: 18pt;
+		}
+
+		.round {
+			display: flex;
+			flex-direction: column;
+			gap: 10px;
+
+			.round-text {
+				color: yellow;
+				font-size: 18pt;
+			}
+		}
+
+		.target-text {
+			color: red;
+			font-size: 28pt;
+		}
+
+		.turn-text {
+			color: red;
+			font-size: 24pt;
+		}
+
+		.cheats {
+			display: flex;
+			flex-direction: column;
+			gap: 10px;
+		}
+	}
+}
