+++ src/number-rogue/util/scene-management.ts
@@ -1,0 +1,8 @@
+export const SceneType = {
+	Home: "Home",
+	LevelSelect: "LevelSelect",
+	DifficultySelect: "DifficultySelect",
+	Game: "Game",
+};
+
+export type SceneType = (typeof SceneType)[keyof typeof SceneType];
