+++ src/number-rogue/util/seed-gen.ts
