+++ src/number-rogue/vite-env.d.ts
